# public/swagger.v1.json
{
  "consumes": [
    "application/json",
    "text/plain"
  ],
  "produces": [
    "application/json",
    "text/html"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This provide API interface to communicate with this Gitea instance.",
    "title": "Gitea API.",
    "termsOfService": "there are no TOS at this moment, use at your own risk we take no responsibility",
    "license": {
      "name": "MIT",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "1.1.1"
  },
  "basePath": "/api/v1",
  "paths": {
    "/markdown": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/html"
        ],
        "operationId": "renderMarkdown",
        "parameters": [
          {
            "description": "Text markdown to render",
            "name": "Text",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Mode to render",
            "name": "Mode",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Context to render",
            "name": "Context",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Is it a wiki page ?",
            "name": "Wiki",
            "in": "body",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MarkdownRender"
          },
          "422": {
            "$ref": "#/responses/validationError"
          }
        }
      }
    },
    "/markdown/raw": {
      "post": {
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "text/html"
        ],
        "operationId": "renderMarkdownRaw",
        "responses": {
          "200": {
            "$ref": "#/responses/MarkdownRender"
          },
          "422": {
            "$ref": "#/responses/validationError"
          }
        }
      }
    },
    "/org/{org}/repos": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "createOrgRepo",
        "parameters": [
          {
            "uniqueItems": true,
            "x-go-name": "Name",
            "description": "Name of the repository to create",
            "name": "name",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "Description",
            "description": "Description of the repository to create",
            "name": "description",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "Private",
            "description": "Is the repository to create private ?",
            "name": "private",
            "in": "body",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "x-go-name": "AutoInit",
            "description": "Init the repository to create ?",
            "name": "auto_init",
            "in": "body",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "x-go-name": "Gitignores",
            "description": "Gitignores to use",
            "name": "gitignores",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "License",
            "description": "License to use",
            "name": "license",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "Readme",
            "description": "Readme of the repository to create",
            "name": "readme",
            "in": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/Repository"
          },
          "403": {
            "$ref": "#/responses/forbidden"
          },
          "422": {
            "$ref": "#/responses/validationError"
          },
          "500": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
    "/repos/search": {
      "get": {
        "produces": [
          "application/json"
        ],
        "operationId": "repoSearch",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Keyword",
            "description": "Keyword to search",
            "name": "q",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "OwnerID",
            "description": "Owner in we search search",
            "name": "uid",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "PageSize",
            "description": "Limit of result\n\nmaximum: setting.ExplorePagingNum",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/SearchResults"
          },
          "500": {
            "$ref": "#/responses/SearchError"
          }
        }
      }
    },
...
    "/user/followers": {
      "get": {
        "produces": [
          "application/json"
        ],
        "operationId": "userCurrentListFollowers",
        "responses": {
          "200": {
            "$ref": "#/responses/UserList"
          },
          "500": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
    "/user/following": {
      "get": {
        "produces": [
          "application/json"
        ],
        "operationId": "userCurrentListFollowing",
        "responses": {
          "200": {
            "$ref": "#/responses/UserList"
          },
          "500": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
...
    "/user/gpg_keys": {
      "get": {
        "produces": [
          "application/json"
        ],
        "operationId": "userCurrentListGPGKeys",
        "responses": {
          "200": {
            "$ref": "#/responses/GPGKeyList"
          },
          "500": {
            "$ref": "#/responses/error"
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "userCurrentPostGPGKey",
        "parameters": [
          {
            "uniqueItems": true,
            "x-go-name": "ArmoredKey",
            "description": "An armored GPG key to add",
            "name": "armored_public_key",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/GPGKey"
          },
          "422": {
            "$ref": "#/responses/validationError"
          },
          "500": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
...
    "/user/keys": {
      "get": {
        "produces": [
          "application/json"
        ],
        "operationId": "userCurrentListKeys",
        "responses": {
          "200": {
            "$ref": "#/responses/PublicKeyList"
          },
          "500": {
            "$ref": "#/responses/error"
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "userCurrentPostKey",
        "parameters": [
          {
            "uniqueItems": true,
            "x-go-name": "Title",
            "description": "Title of the key to add",
            "name": "title",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "uniqueItems": true,
            "x-go-name": "Key",
            "description": "An armored SSH key to add",
            "name": "key",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/PublicKey"
          },
          "422": {
            "$ref": "#/responses/validationError"
          },
          "500": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
...
    "/user/repos": {
      "get": {
        "produces": [
          "application/json"
        ],
        "operationId": "userCurrentListRepos",
        "responses": {
          "200": {
            "$ref": "#/responses/RepositoryList"
          },
          "500": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
    "/user/starred": {
      "get": {
        "produces": [
          "application/json"
        ],
        "operationId": "userCurrentListStarred",
        "responses": {
          "200": {
            "$ref": "#/responses/RepositoryList"
          },
          "500": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
...
    "/user/subscriptions": {
      "get": {
        "produces": [
          "application/json"
        ],
        "operationId": "userCurrentListSubscriptions",
        "responses": {
          "200": {
            "$ref": "#/responses/RepositoryList"
          },
          "500": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
...
    "/users/search": {
      "get": {
        "produces": [
          "application/json"
        ],
        "operationId": "userSearch",
        "responses": {
          "200": {
            "$ref": "#/responses/UserList"
          },
          "500": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
    "/users/{username}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "operationId": "userGet",
        "responses": {
          "200": {
            "$ref": "#/responses/User"
          },
          "404": {
            "$ref": "#/responses/notFound"
          },
          "500": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "GPGKeyEmail": {
      "description": "GPGKeyEmail a email attache to a GPGKey",
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "verified": {
          "type": "boolean",
          "x-go-name": "Verified"
        }
      },
      "x-go-package": "code.gitea.io/gitea/vendor/code.gitea.io/sdk/gitea"
    },
    "Permission": {
      "type": "object",
      "title": "Permission represents a API permission.",
      "properties": {
        "admin": {
          "type": "boolean",
          "x-go-name": "Admin"
        },
        "pull": {
          "type": "boolean",
          "x-go-name": "Pull"
        },
        "push": {
          "type": "boolean",
          "x-go-name": "Push"
        }
      },
      "x-go-package": "code.gitea.io/gitea/vendor/code.gitea.io/sdk/gitea"
    },
    "Repository": {
      "type": "object",
      "title": "Repository represents a API repository.",
      "properties": {
        "clone_url": {
          "type": "string",
          "x-go-name": "CloneURL"
        },
        "created_at": {
          "x-go-name": "Created"
        },
        "default_branch": {
          "type": "string",
          "x-go-name": "DefaultBranch"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "fork": {
          "type": "boolean",
          "x-go-name": "Fork"
        },
        "forks_count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Forks"
        },
        "full_name": {
          "type": "string",
          "x-go-name": "FullName"
        },
        "html_url": {
          "type": "string",
          "x-go-name": "HTMLURL"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "mirror": {
          "type": "boolean",
          "x-go-name": "Mirror"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "open_issues_count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "OpenIssues"
        },
        "owner": {
          "x-go-name": "Owner",
          "$ref": "#/definitions/User"
        },
        "permissions": {
          "x-go-name": "Permissions",
          "$ref": "#/definitions/Permission"
        },
        "private": {
          "type": "boolean",
          "x-go-name": "Private"
        },
        "ssh_url": {
          "type": "string",
          "x-go-name": "SSHURL"
        },
        "stars_count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Stars"
        },
        "updated_at": {
          "x-go-name": "Updated"
        },
        "watchers_count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Watchers"
        },
        "website": {
          "type": "string",
          "x-go-name": "Website"
        }
      },
      "x-go-package": "code.gitea.io/gitea/vendor/code.gitea.io/sdk/gitea"
    },
    "User": {
      "type": "object",
      "title": "User represents a API user.",
      "properties": {
        "avatar_url": {
          "type": "string",
          "x-go-name": "AvatarURL"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "full_name": {
          "type": "string",
          "x-go-name": "FullName"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "login": {
          "type": "string",
          "x-go-name": "UserName"
        }
      },
      "x-go-package": "code.gitea.io/gitea/vendor/code.gitea.io/sdk/gitea"
    }
  },
  "responses": {
...
    "GPGKey": {
      "description": "GPGKey a user GPG key to sign commit and tag in repository",
      "headers": {
        "can_certify": {
          "type": "boolean"
        },
        "can_encrypt_comms": {
          "type": "boolean"
        },
        "can_encrypt_storage": {
          "type": "boolean"
        },
        "can_sign": {
          "type": "boolean"
        },
        "created_at": {},
        "emails": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GPGKeyEmail"
          }
        },
        "expires_at": {},
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "key_id": {
          "type": "string"
        },
        "primary_key_id": {
          "type": "string"
        },
        "public_key": {
          "type": "string"
        },
        "subkeys": {
          "type": "array",
          "items": {
            "type": "object"
          }
        }
      }
    },
    "GPGKeyList": {
      "description": "GPGKeyList represents a list of GPGKey"
    },
    "MarkdownRender": {
      "description": "MarkdownRender is a rendered markdown document"
    },
    "PublicKey": {
      "description": "PublicKey publickey is a user key to push code to repository",
      "headers": {
        "created_at": {},
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "key": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
    "PublicKeyList": {
      "description": "PublicKeyList represents a list of PublicKey"
    },
    "Repository": {
      "description": "Repository represents a API repository.",
      "schema": {
        "$ref": "#/definitions/Permission"
      },
      "headers": {
        "clone_url": {
          "type": "string"
        },
        "created_at": {},
        "default_branch": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "fork": {
          "type": "boolean"
        },
        "forks_count": {
          "type": "integer",
          "format": "int64"
        },
        "full_name": {
          "type": "string"
        },
        "html_url": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "mirror": {
          "type": "boolean"
        },
        "name": {
          "type": "string"
        },
        "open_issues_count": {
          "type": "integer",
          "format": "int64"
        },
        "owner": {},
        "permissions": {},
        "private": {
          "type": "boolean"
        },
        "ssh_url": {
          "type": "string"
        },
        "stars_count": {
          "type": "integer",
          "format": "int64"
        },
        "updated_at": {},
        "watchers_count": {
          "type": "integer",
          "format": "int64"
        },
        "website": {
          "type": "string"
        }
      }
    },
    "RepositoryList": {
      "description": "RepositoryList represents a list of API repository."
    },
    "SearchError": {
      "description": "SearchError error of failing search",
      "headers": {
        "error": {
          "type": "string"
        },
        "ok": {
          "type": "boolean"
        }
      }
    },
    "SearchResults": {
      "description": "SearchResults results of search",
      "headers": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Repository"
          }
        },
        "ok": {
          "type": "boolean"
        }
      }
    },
...
    "User": {
      "description": "User represents a API user.",
      "headers": {
        "avatar_url": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "full_name": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "login": {
          "type": "string"
        }
      }
    },
    "UserList": {
      "description": "UserList represents a list of API user."
    },
...
    "error": {
      "description": "APIError is error format response",
      "headers": {
        "message": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
    "forbidden": {
      "description": "APIForbiddenError is a forbidden error response",
      "headers": {
        "message": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
    "notFound": {
      "description": "APINotFound is a not found empty response"
    },
    "validationError": {
      "description": "APIValidationError is error format response related to input validation",
      "headers": {
        "message": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    }
  }
}